small tweaks to grafana workflows and moved the grafana package
