+++ .github/workflows/dev-deploy-grafana.yml
@@ -36,7 +36,7 @@ jobs:
 
             infisical export \
               --token=$INFISICAL_TOKEN \
-              --env=master \
+              --env=dev \
               --format=dotenv \
               --path="/api" \
               --projectId='${{ secrets.INFISICAL_PROJECT_ID }}' \
